Edit Login.html, added Login.css & template.html
Edited Login.html to have form features
Still need to work on the Design of the form by editing the
box size of username and password box.

template.html includes a template code that I found online.

Login.css is a file that I will use to input all the styles in.
This is to save space in Login.html file.